[broadcast] Fixes bug when cloning multiple times
Before this commit, in Chrome, when cloning multiple times a graph, the
oldest clones start losing their WebGL contexts at some point. The
reason is that the contexts are actually shared along tabs/windows, as
long as they allow referencing each other (which kind of makes some
sense I guess?).

The fix is to add by default "noopener" to the window.open call.