[gephi-lite] Improves search in data table
This commit just removes the sigmaGraph dependency on the matchingItems
and otherItems useMemo generation, since when a new node is created, the
filtered graph knows it before the sigma graph.

This commit partially solves #251, since there is no longer an error -
but the new node still does not appear in the list (because of some
other indexation issues).