[gephi-lite] Fix extractGraphFromFile
When opening a reomte file, some will have query params, such as a pre-signed url to a cloud hosted storage bucket.
Ex: https://www.example.com/graph.gexf?token=XXX

But url like https://www.example.com/index.php?file=graph.gexf are also
valid, as well as https://www.example.com/xzy (which can be a shared
link with a valid file).

So computing the file format from the filename is not a the good way to
do it, we need to parse the file content and check what it is.