[gephhi-lite] add 'new version available' on welcome modale
When deploying the a new version, the CI produce a versions.json file at
the root of website.

GL calls this endpoint (or not can be configured in gephi lite config with
some env variable), and if in the versions.json file tehre is highest
version that the app one, a text is displayed on the Welcome modal.

NOTE: The versions.json file has been design to be able to handle
multiple versions (it's an array), byt due to CI process I can only
override it (without reading it at first). SO the file contains only the
latest deployed version, which is the lastest one (I think we wil never
push a fix on an old version)